Standalone collection & Stereotype Annotations
Created a Standalone collection of type LinkedList and injecting it.  Created an object using the @Component annotation at a class level(creates an object of that class and names it the same but with camel case).  I used the @Scope annotation at the class level to create different instances of an object each time it has been read/used in the application. Here there is also a demonstration of using the @Value annotation to inject Primitives types and Collection types, and the @Autowired to inject Object types.